Started refactoring
Refactored the crud services. Made a rest apis, rest. Made a new controller for the api-key. Made the dtos more unified where it could be. Started working on the custom annotations.

+++ src/main/java/com/example/ehe_server/service/log/LoggingService.java
@@ -45,7 +45,7 @@ public class LoggingService implements LoggingServiceInterface {
 
         // Set user if it's a human user
         if (userContextService.isHumanUser()) {
-            Optional<User> user = userRepository.findById(userContextService.getCurrentUserId().intValue());
+            Optional<User> user = userRepository.findById(userContextService.getCurrentUserId());
             if (user.isPresent()) {
                 log.setUser(user.get());
             }
@@ -69,7 +69,7 @@ public class LoggingService implements LoggingServiceInterface {
 
         // Set user if it's a human user
         if (userContextService.isHumanUser()) {
-            Optional<User> user = userRepository.findById(userContextService.getCurrentUserId().intValue());
+            Optional<User> user = userRepository.findById(userContextService.getCurrentUserId());
             if (user.isPresent()) {
                 errorLog.setUser(user.get());
             }
